+++ account.php
@@ -275,7 +275,7 @@ $flashSuccess = flash('success');
     <div id="cookie-notice-root"></div>
     <div id="age-gate-root"></div>
 
-    <script>
+    <script<?= nonce_attr(); ?>>
         window.__VIDEW__ = <?= page_bootstrap(default_bootstrap_payload('account')); ?>;
     </script>
     <?= gui_runtime_tags(); ?>