+++ become-creator.php
@@ -163,7 +163,7 @@ $applicationStatus = (string) ($application['status'] ?? '');
     <div id="cookie-notice-root"></div>
     <div id="age-gate-root"></div>
 
-    <script>
+    <script<?= nonce_attr(); ?>>
         window.__VIDEW__ = <?= page_bootstrap(default_bootstrap_payload('become-creator')); ?>;
     </script>
     <?= gui_runtime_tags(); ?>