+++ browse.php
@@ -192,7 +192,7 @@ $bootPayload = default_bootstrap_payload('browse', [
     <div id="cookie-notice-root"></div>
     <div id="age-gate-root"></div>
 
-    <script>
+    <script<?= nonce_attr(); ?>>
         window.__VIDEW__ = <?= page_bootstrap($bootPayload); ?>;
     </script>
     <?= gui_runtime_tags(); ?>