+++ channel.php
@@ -136,7 +136,7 @@ $channelName = $creator ? creator_public_name($creator) : 'Channel';
     <div id="cookie-notice-root"></div>
     <div id="age-gate-root"></div>
 
-    <script>
+    <script<?= nonce_attr(); ?>>
         window.__VIDEW__ = <?= page_bootstrap(default_bootstrap_payload('channel')); ?>;
     </script>
     <?= gui_runtime_tags(); ?>