+++ mfa-challenge.php
@@ -71,7 +71,7 @@ if (is_post_request()) {
     <div id="cookie-notice-root"></div>
     <div id="age-gate-root"></div>
 
-    <script>
+    <script<?= nonce_attr(); ?>>
         window.__VIDEW__ = <?= page_bootstrap(default_bootstrap_payload('mfa-challenge')); ?>;
     </script>
     <?= gui_runtime_tags(); ?>